+++ src/pages/stashes/stashes.tsx
@@ -8,8 +8,9 @@ import { useDimensions } from '../../components/hooks/useDimensions';
 import { useTruncationMode } from '../../components/hooks/useTruncationMode';
 import { useEffect } from 'react';
 import { useErrorCatcher } from '../../components/hooks/useErrorCatcher';
-import { QuickPick, useModal } from '../../components/modal';
+import { ModalInput, QuickPick, useModal } from '../../components/modal';
 import { useNav } from '../../components/navigation';
+import { revParse } from '../../utils/git';
 
 type StashesProps = {
   setStash: Dispatch<SetStateAction<string | undefined>>;
@@ -20,7 +21,7 @@ export function Stashes({ setStash }: StashesProps) {
   const { mode } = useTruncationMode();
   const { attempt } = useErrorCatcher();
   const { setModal, open } = useModal();
-  const { activeSection } = useNav();
+  const { activeSection, isLocked } = useNav();
 
   const { value: allStashes = [], refresh } = useResolved(getStashList);
   const {
@@ -40,7 +41,24 @@ export function Stashes({ setStash }: StashesProps) {
     refresh();
   };
 
+  const reword = async () => {
+    const stash = `stash@{${selectedIndex}}`;
+    const ref = await revParse(stash);
+    setModal(
+      <ModalInput
+        title={`reword ${stash}`}
+        handleSubmit={async (msg) => {
+          await execAsync(`git stash drop ${stash}`);
+          await execAsync(`git stash store ${ref} -m "${msg}"`);
+          refresh();
+        }}
+      />
+    );
+    open();
+  };
+
   useInput((input, key) => {
+    if (isLocked) return;
     if (input === 's') {
       setModal(
         <QuickPick
@@ -64,6 +82,8 @@ export function Stashes({ setStash }: StashesProps) {
       attempt(cmdBuilder('pop'));
     } else if (input === 'd') {
       attempt(cmdBuilder('drop'));
+    } else if (input === 'r') {
+      attempt(reword);
     }
 
     if (activeSection !== 'Stashes') return;